+++ modules/robot/src/wireframe-simulator/core/vpMyio.cpp
@@ -106,16 +106,16 @@ void fscanf_string (char **str)
 	if (lex () != T_STRING)
 		lexerr ("start", "string expected", NULL);
 	if (*str == NULL)
-		*str = (char *) malloc ((mylength + 1) * sizeof (char));
+    *str = (char *) malloc ((size_t)(mylength + 1) * sizeof (char));
   else
-    *str = (char *) realloc (*str, (mylength + 1) * sizeof (char));
+    *str = (char *) realloc (*str, (size_t)(mylength + 1) * sizeof (char));
 
   if (*str == NULL) {
     printf("Unable to read the string: bad memory allocation");
     return;
   }
 
-	strncpy (*str, mytext, mylength);
+  strncpy (*str, mytext, (size_t)mylength);
 }
 
 /*
