+++ modules/robot/src/wireframe-simulator/vpWireFrameSimulator.cpp
@@ -428,7 +428,7 @@ vpWireFrameSimulator::display_scene(Matrix mat, Bound_scene &sc, const vpImage<v
 
       set_Bound_face_display (clip, b); //regarde si is_visible
 
-      point_3D_2D (clip->point.ptr, clip->point.nbr,I.getWidth(),I.getHeight(),point2i);
+      point_3D_2D (clip->point.ptr, clip->point.nbr,(int)I.getWidth(),(int)I.getHeight(),point2i);
       for (; fp < fend; fp++)
       {
         if (fp->is_visible)
@@ -478,7 +478,7 @@ vpWireFrameSimulator::display_scene(Matrix mat, Bound_scene &sc, const vpImage<u
 
       set_Bound_face_display (clip, b); //regarde si is_visible
 
-      point_3D_2D (clip->point.ptr, clip->point.nbr,I.getWidth(),I.getHeight(),point2i);
+      point_3D_2D (clip->point.ptr, clip->point.nbr,(int)I.getWidth(),(int)I.getHeight(),point2i);
       for (; fp < fend; fp++)
       {
         if (fp->is_visible)
